requirements.txt
pytest
flask
selenium
webdriver-manager
allure-pytest